# design/api/openapi.yaml
openapi: 3.1.0
info:
  title: Fiscal Note Workbench API
  version: 1.0.0
  description: |
    REST API for the DOR fiscal note proof of concept. Paths are grouped by module:
    identity, bills, search, notes, templates, workflow, notifications, reference, admin.
    Detailed request and response shapes are specified in design/research/*.md; this file
    is the route catalog and the shared schema skeleton the implementation must generate from
    or conform to. All paths are served under /api/v1.
servers:
  - url: /api/v1
security:
  - session: []
  - bearer: []
...
paths:
...
  /search:
    get:
      tags: [search]
      summary: Text search with direct-hit resolution, filters, facets (research/search.md section 6.1)
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: biennium, in: query, schema: { type: string } }
        - { name: chamber, in: query, schema: { type: string, enum: [H, S] } }
        - { name: type, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: committee, in: query, schema: { type: string } }
        - { name: sponsor, in: query, schema: { type: string } }
        - { name: has_fiscal_note, in: query, schema: { type: boolean } }
        - { name: fiscal_note_status, in: query, schema: { type: string } }
        - { name: doc_type, in: query, schema: { type: string } }
        - { name: rcw, in: query, schema: { type: string } }
        - { name: version_code, in: query, schema: { type: string } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
        - { name: assigned_to_me, in: query, schema: { type: boolean } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: sort, in: query, schema: { type: string, enum: [relevance, date, bill_number] } }
      responses:
        "200": { description: SearchResponse, content: { application/json: { schema: { $ref: "#/components/schemas/SearchResponse" } } } }
        "503": { description: search_unavailable }
  /search/suggest:
    get:
      tags: [search]
      summary: Search-as-you-type suggestions
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 1 } }
        - { name: biennium, in: query, schema: { type: string } }
        - { name: size, in: query, schema: { type: integer, default: 8, maximum: 10 } }
      responses:
        "200": { description: Suggestions, content: { application/json: { schema: { $ref: "#/components/schemas/SuggestResponse" } } } }
  /search/reindex:
    post:
      tags: [admin]
      summary: Rebuild or refresh indices (admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                scope: { type: string, enum: [full, incremental, bill] }
                bill_keys: { type: array, items: { type: string } }
                indices: { type: array, items: { type: string } }
      responses:
        "202": { description: Job accepted, content: { application/json: { schema: { $ref: "#/components/schemas/JobAccepted" } } } }

  # ---------------- notes ----------------
...
  /templates:
    get:
      tags: [templates]
      summary: List templates
      parameters:
        - { name: mode, in: query, schema: { type: string, enum: [limited, full] } }
        - { name: kind, in: query, schema: { type: string, enum: [document, snippet] } }
        - { name: taxType, in: query, schema: { type: string } }
        - { name: impactType, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200": { description: List, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/TemplateSummary" } } } } }
  /templates/{id}:
    get:
      tags: [templates]
      summary: Template with HTML body
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Template, content: { application/json: { schema: { $ref: "#/components/schemas/Template" } } } }
    put:
      tags: [templates]
      summary: Update a template (template_editor role); creates a new template version
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/Template" } } } }
      responses: { "200": { description: Updated } }
  /templates/{id}/preview:
    get:
      tags: [templates]
      summary: Template rendered with a note's token context
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: noteId, in: query, schema: { type: string } }
      responses:
        "200": { description: HTML, content: { text/html: { schema: { type: string } } } }

  # ---------------- workflow ----------------
...
components:
  securitySchemes:
    session: { type: apiKey, in: cookie, name: session }
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
...
  schemas:
...
    SearchResponse:
      type: object
      description: research/search.md section 6.1
      properties:
        query: { type: string }
        parsed: { type: object }
        direct: { type: object, nullable: true }
        hits: { type: array, items: { type: object } }
        facets: { type: object }
        page: { type: integer }
        size: { type: integer }
        total: { type: integer }
        took_ms: { type: integer }
        backend: { type: string }
    SuggestResponse:
      type: object
      properties:
        query: { type: string }
        reference: { type: object }
        suggestions: { type: array, items: { type: object } }
...
    TemplateSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        kind: { type: string, enum: [document, snippet] }
        mode: { type: string, enum: [limited, full] }
        version: { type: integer }
        description: { type: string }
        tags: { type: array, items: { type: string } }
        parts: { type: array, items: { type: string } }
        slots: { type: array, items: { type: object } }
        tokens: { type: array, items: { type: string } }
    Template:
      allOf:
        - { $ref: "#/components/schemas/TemplateSummary" }
        - type: object
          properties:
            html: { type: string }
            etag: { type: string }
...
    JobAccepted:
      type: object
      properties:
        job_id: { type: string }
        status_url: { type: string }